Add sql interface python script for SkyhookDM
This is a very early version of the SQL interface for SkyhookDM. This initial commit only supports SELECT-FROM queries. Expect this to be extended in the upcoming week to support WHERE clauses, more options, etc. 

Upcoming hot fix to update SkyhookDM to be most recent version and handle any name mismatches from recent updates to options in recent SkyhookDM commits.